⚙️ Refactor user page folder
