+++ client/src/features/artist/components/Artist.module.scss
