+++ client/src/features/artist/components/Artist.tsx
