+++ client/src/features/artist/components/ArtistTable.tsx
