+++ client/src/pages/ArtistPage.tsx
@@ -1,4 +1,4 @@
-import Artist from '../features/artist/componens/Artist.tsx';
+import Artist from '../features/artist/components/Artist.tsx';
 
 const ArtistPage = () => {
   return <Artist />;
